requirements.txt
opencv-python
dlib
numpy
pystray
Pillow
win10toast